Backend/requirements.txt
fastapi
uvicorn
groq
python-multipart
python-dotenv
PyPDF2 